brought back the new logic to wrapper.js
